feat!: revise combinator naming and switch counts from uint to int (#152)
Closes #151

- `Prefixed` -> `Preceded` (arguments flipped to match call order)
- `Suffixed` -> `Terminated`
- `Any` -> `IsA`
- `Not` -> `IsNot`
- Fill removed (was an exact alias of Repeat)
- count parameters switched from uint to int
- `Named` added, for building a Predicate from a plain function

+++ predicate_test.go
@@ -513,7 +513,7 @@ func TestWhileNMultiByteBoundary(t *testing.T) {
 
 	var rangedErr chomp.RangedParserError
 	require.ErrorAs(t, err, &rangedErr)
-	assert.Equal(t, uint(5), rangedErr.Exec.Count, "count must report runes, not bytes")
+	assert.Equal(t, 5, rangedErr.Exec.Count, "count must report runes, not bytes")
 }
 
 func TestWhileNMMultiByteBoundary(t *testing.T) {
@@ -535,7 +535,7 @@ func TestWhileNotNMultiByteBoundary(t *testing.T) {
 
 	var rangedErr chomp.RangedParserError
 	require.ErrorAs(t, err, &rangedErr)
-	assert.Equal(t, uint(5), rangedErr.Exec.Count, "count must report runes, not bytes")
+	assert.Equal(t, 5, rangedErr.Exec.Count, "count must report runes, not bytes")
 }
 
 func TestWhileNotNMMultiByteBoundary(t *testing.T) {
